Make DSI members consistent in failover section
The section about failover is inconsistent with respect to the
roles of the DSI cluster members. For example, first it's shown
that DSI cluster member with index 0 is the primary. In the next
command (which corresponds to the same DSI at the same time as
the previous command), it is shown that the cluster member with
index 1 is the primary. This commit fixes that by ensuring that DSI
cluster members roles are consistent.